Add markdown source of manual
This packages the manual in the app, and loads an instance of it during
the first run.

The manual content is here in a series of markdown pages, and a set of
accompanying images. Hopefully this makes it easy to maintain going
forward.

In order to resolve attachments and links properly, there are a couple
of things to bear in mind when authoring the manual content:

- Links to images and other pages should be relative! If copying the
  content from an existing site, make sure to drop any leading host
  part, like `https://books.example.com`.
- Links from one page to another usually depend on their ID. But the IDs
  of the newly-built book may not match where the source was written. To
  solve this, make sure the slug part matches the title of the book. The
  code will use that to find the right page, and then update the IDs in
  the content. For example, if you originally had a link like:

      /12/manual/83/introduction

   Then as long as there is a page called `Introduction`, the script can
   fix this up to use the correct IDs, possibly resulting in a link
   like:

      /1/manual/4/introduction

To test out any future changes to this content, you can build new copies
of the manual from a dev console:

```ruby
DemoContent.create_manual(User.first)
```